refactor: address Aditya's PR feedback
* Extract dialog into reusable EditBranchDialog component (project list,
  dashboard list can drive it via the open prop)
* Use existing Tabs primitive (Tabs/TabsList/TabsTrigger/TabsContent
  with data-[state=active] for the segmented styling — drops the
  custom :global() CSS and dark-mode-specific overrides)
* Use forms/Input for the new-branch field — built-in label, errors,
  styling
* Drop Dialog.Content overrides; inherit defaults
* Replace handcrafted styles with Tailwind utilities throughout
* handleTabChange centralizes per-tab state cleanup
